chore: created spid_cie_oidc.authority.validators and moved validate_entity_configuration from models
